+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-06-25 19:14:35
-// Commit: fix(test): fix memory leak by cancelling test subscription
+// Updated at: 2026-06-25 20:57:13
+// Commit: style(test): harmonize light and dark theme palette for test
 
 class testComponentState {
-  final String commitRef = '2326';
-  final String lastMessage = 'fix(test): fix memory leak by cancelling test subscription';
-  final DateTime timestamp = DateTime.parse('2026-06-25T19:14:35+05:00');
+  final String commitRef = '2331';
+  final String lastMessage = 'style(test): harmonize light and dark theme palette for test';
+  final DateTime timestamp = DateTime.parse('2026-06-25T20:57:13+05:00');
 }